# backend/public/openapi.yaml
openapi: 3.0.3
info:
  title: 'Secure Drop API'
  description: |
    Secure Drop allows you to securely share sensitive information (like passwords or API keys) that self-destructs after being viewed.

    ## Features
    - **Burn on Read**: Secrets are deleted after being viewed
    - **Time-based Expiration**: Optional TTL for secrets
    - **View Limits**: Control how many times a secret can be viewed
    - **Encryption at Rest**: All secrets are encrypted in the database
    - **Secure IDs**: Non-sequential IDs prevent URL enumeration attacks
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://secure-drop.localhost
    description: Local development server
...
paths:
...
  /api/v1/secrets:
    post:
      tags:
        - Secrets
      summary: Create a new secret
      description: Creates a new encrypted secret that will self-destruct after being viewed or after expiration
      operationId: createSecret
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: The secret text to encrypt
                  example: my-api-key-12345
                  maxLength: 10000
                ttlSeconds:
                  type: integer
                  description: Time to live in seconds (10 to 2592000, ~30 days)
                  example: 3600
                  minimum: 10
                  maximum: 2592000
                maxViews:
                  type: integer
                  description: Maximum number of views allowed (1 to 100)
                  example: 1
                  minimum: 1
                  maximum: 100
                  default: 1
      responses:
        '201':
          description: Secret created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        description: The unique ID of the secret
                        example: V1StGXR8_Z5jdHi6B-myT
                      url:
                        type: string
                        format: uri
                        description: The URL to view the secret
                        example: http://secure-drop-ui.localhost/secrets/V1StGXR8_Z5jdHi6B-myT
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...
components:
  schemas:
    ValidationError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: The given data was invalid.
            details:
              type: object
              example:
                text:
                  - The secret text is required.
...